Split project to different files, and add draw function as a method to the world struct
